# identity-ms/identity-ms-api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.2.0-SNAPSHOT",
    "title" : "WSO2UserManagerAPI"
  },
  "basePath" : "/",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/admin/bulk/permissions" : {
      "patch" : {
        "tags" : [ "Bulk API" ],
        "summary" : "Update permissions in bulk mode",
        "description" : "",
        "operationId" : "bulkUpdatePermissions",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/RolePermissionBulkDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "201" : {
            "description" : "Valid user is created",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "409" : {
            "description" : "errors",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
    "/admin/bulk/roles" : {
      "post" : {
        "tags" : [ "Bulk API" ],
        "summary" : "Create roles in bulk mode",
        "description" : "",
        "operationId" : "bulkCreateRoles",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/RoleCreateBulkDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "201" : {
            "description" : "Valid roles are created",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "409" : {
            "description" : "errors",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Bulk API" ],
        "summary" : "Delete roles in bulk mode",
        "description" : "",
        "operationId" : "bulkDeleteRoles",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/RoleDeleteBulkDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "201" : {
            "description" : "Roles are deleted",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "409" : {
            "description" : "errors",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "Bulk API" ],
        "summary" : "Update roles in bulk mode",
        "description" : "",
        "operationId" : "bulkUpdateRoles",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/RoleUpdateBulkDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "201" : {
            "description" : "Valid user is created",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "409" : {
            "description" : "errors",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
    "/admin/bulk/users" : {
      "post" : {
        "tags" : [ "Bulk API" ],
        "summary" : "Create users in bulk mode",
        "description" : "",
        "operationId" : "bulkCreateUsers",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/UserCreateBulkDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "201" : {
            "description" : "Valid users are created",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "409" : {
            "description" : "errors",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Bulk API" ],
        "summary" : "Delete users in bulk mode",
        "description" : "",
        "operationId" : "bulkDeleteUsers",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/UserDeleteBulkDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "201" : {
            "description" : "Users are deleted",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          },
          "409" : {
            "description" : "errors",
            "schema" : {
              "$ref" : "#/definitions/ResponseDTO"
            }
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
    "/admin/roles" : {
      "post" : {
        "tags" : [ "Role API" ],
        "summary" : "Create new Role",
        "description" : "",
        "operationId" : "createRole",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Role"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Valid Role is created"
          },
          "404" : {
            "description" : "Role is not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Role API" ],
        "summary" : "Delete a given role name",
        "description" : "",
        "operationId" : "deleteRole",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Role is deleted"
          },
          "404" : {
            "description" : "Valid Role is not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "Authorization API" ],
        "summary" : "This function clears all authorizations of the role",
        "description" : "",
        "operationId" : "clearAllRoleAuthorization",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : ""
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
    "/admin/roles/_rename" : {
      "put" : {
        "tags" : [ "Role API" ],
        "summary" : "Rename Role",
        "description" : "",
        "operationId" : "renameRole",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Role Rename"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Valid Role is created"
          },
          "404" : {
            "description" : "Role is not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
    "/admin/roles/_update" : {
      "put" : {
        "tags" : [ "Role API" ],
        "summary" : "Update Role",
        "description" : "",
        "operationId" : "updateRole",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/RoleUpdateDTO"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Valid Role is created"
          },
          "404" : {
            "description" : "Role is not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
    "/roles" : {
      "get" : {
        "tags" : [ "Role API" ],
        "summary" : "Get a list of all the roles created in the system",
        "description" : "",
        "operationId" : "getRoles",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Valid Roles are found",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "404" : {
            "description" : "Valid Roles are not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
    "/roles/users" : {
      "get" : {
        "tags" : [ "Role API" ],
        "summary" : "Retrieve a list of all users belonging to a role",
        "description" : "",
        "operationId" : "getUsersOfRole",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Valid Role are found",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "404" : {
            "description" : "Valid Role are not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
    "/users" : {
      "get" : {
        "tags" : [ "User API" ],
        "summary" : "Retrieve a list of all users",
        "description" : "",
        "operationId" : "getUsers",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "fetch",
          "in" : "query",
          "description" : "Returns users according to the filter with or without claims",
          "required" : true,
          "type" : "string",
          "enum" : [ "eager" ]
        }, {
          "name" : "filter",
          "in" : "query",
          "description" : "Filter expression for filtering",
          "required" : true,
          "type" : "string",
          "x-example" : "pete*, *mith"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "Specifies the desired maximum number of query results",
          "required" : true,
          "type" : "integer",
          "format" : "int32",
          "x-example" : 10
        } ],
        "responses" : {
          "200" : {
            "description" : "Valid users are returned",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/User Fetch"
              }
            }
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
    "/users/roles" : {
      "get" : {
        "tags" : [ "User API" ],
        "summary" : "Get the list of roles that a particular user belongs to",
        "description" : "",
        "operationId" : "getRolesOfUser",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "username",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Valid user is updated",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    }
  },
  "securityDefinitions" : {
    "X-Auth-Token" : {
      "type" : "apiKey",
      "name" : "X-Auth-Token",
      "in" : "header"
    }
  },
  "definitions" : {
    "KOMessageDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "readOnly" : true
        },
        "message" : {
          "type" : "string",
          "readOnly" : true
        }
      }
    },
    "ResponseDTO" : {
      "type" : "object",
      "properties" : {
        "ko" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/KOMessageDTO"
          }
        }
      }
    },
    "Role" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "readOnly" : true
        },
        "newUsers" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        },
        "permissions" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
    "RoleCreateBulkDTO" : {
      "type" : "object",
      "required" : [ "newRoles" ],
      "properties" : {
        "newRoles" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/Role"
          }
        }
      }
    },
    "RoleDeleteBulkDTO" : {
      "type" : "object",
      "properties" : {
        "deletedRoles" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    },
    "RolePermissionBulkDTO" : {
      "type" : "object",
      "required" : [ "updatedRoles" ],
      "properties" : {
        "updatedRoles" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/RolePermissionItemBulkDTO"
          }
        }
      }
    },
    "RolePermissionItemBulkDTO" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "readOnly" : true
        },
        "authorizePermissions" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        },
        "denyPermissions" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    },
    "RoleUpdateBulkDTO" : {
      "type" : "object",
      "required" : [ "updatedRoles" ],
      "properties" : {
        "updatedRoles" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/RoleUpdateItemBulkDTO"
          }
        }
      }
    },
    "RoleUpdateDTO" : {
      "type" : "object",
      "properties" : {
        "newUsers" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        },
        "deletedUsers" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    },
    "RoleUpdateItemBulkDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "readOnly" : true
        },
        "newUsers" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        },
        "deletedUsers" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
    "UserCreateBulkDTO" : {
      "type" : "object",
      "properties" : {
        "newUsers" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "$ref" : "#/definitions/User Create"
          }
        }
      }
    },
    "UserDeleteBulkDTO" : {
      "type" : "object",
      "properties" : {
        "deletedUsers" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    }
  }
}